# [java-runtime] Create a .jar to contain the runtime classes