--- a/Documentation/guides/messages/xa0002.md
+++ b/Documentation/guides/messages/xa0002.md
@@ -1,0 +1,5 @@
+# Compiler Error XA0002
+
+This error means the build system could not locate the 
+`mono.android.jar` file. This probably means your installation
+is corrupt in some way. Please reinstall Xamarin.Android.