--- a/build-tools/scripts/JavaCallableWrappers.targets
+++ b/build-tools/scripts/JavaCallableWrappers.targets
@@ -29,9 +29,10 @@
       <_D>-d "$(IntermediateOutputPath)jcw\bin"</_D>
       <_AndroidJar>"$(AndroidToolchainDirectory)\sdk\platforms\android-$(AndroidPlatformId)\android.jar"</_AndroidJar>
       <_MonoAndroidJar>$(OutputPath)mono.android.jar</_MonoAndroidJar>
+      <_MonoAndroidRuntimeJar>$(OutputPath)..\..\..\xbuild\Xamarin\Android\java_runtime.jar</_MonoAndroidRuntimeJar>
     </PropertyGroup>
     <Exec
-        Command="&quot;$(JavaCPath)&quot; $(_Target) $(_D) -bootclasspath $(_AndroidJar)$(PathSeparator)&quot;$(_MonoAndroidJar)&quot; @$(IntermediateOutputPath)jcw\classes.txt"
+        Command="&quot;$(JavaCPath)&quot; $(_Target) $(_D) -bootclasspath $(_AndroidJar)$(PathSeparator)&quot;$(_MonoAndroidJar)&quot;$(PathSeparator)&quot;$(_MonoAndroidRuntimeJar)&quot; @$(IntermediateOutputPath)jcw\classes.txt"
     />
     <Exec
         Condition="Exists('$(_MonoAndroidJar)')"