--- a/src/Xamarin.Android.Build.Tasks/Tasks/BuildApk.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/BuildApk.cs
@@ -48,7 +48,7 @@ namespace Xamarin.Android.Tasks
 		public ITaskItem[] TypeMappings { get; set; }
 
 		[Required]
-		public string [] DalvikClasses { get; set; }
+		public ITaskItem [] DalvikClasses { get; set; }
 
 		[Required]
 		public string SupportedAbis { get; set; }
@@ -105,7 +105,10 @@ namespace Xamarin.Android.Tasks
 						Assembly.GetExecutingAssembly ().GetManifestResourceStream ("NOTICE.txt"));
 
 				// Add classes.dx
-				apk.Archive.AddFiles (DalvikClasses, useFileDirectories: false);
+				foreach (var dex in DalvikClasses) {
+					string apkName = dex.GetMetadata ("ApkName");
+					apk.Archive.AddFile (dex.ItemSpec, string.IsNullOrWhiteSpace (apkName) ? Path.GetFileName (dex.ItemSpec) : apkName);
+				}
 
 				if (EmbedAssemblies && !BundleAssemblies)
 					AddAssemblies (apk);