--- a/src/Xamarin.Android.Build.Tasks/Tasks/GeneratePackageManagerJava.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/GeneratePackageManagerJava.cs
@@ -77,16 +77,10 @@ namespace Xamarin.Android.Tasks
 
 			using (var stream = new MemoryStream ())
 			using (var pkgmgr = new StreamWriter (stream)) {
-				// Write the boilerplate from the MonoPackageManager.java resource
-				using (var template = new StreamReader (Assembly.GetExecutingAssembly ().GetManifestResourceStream ("MonoPackageManager.java"))) {
-					string line;
-					while ((line = template.ReadLine ()) != null) {
-						pkgmgr.WriteLine (line);
-					}
-				}
+				pkgmgr.WriteLine ("package mono;");
 
 				// Write all the user assemblies
-				pkgmgr.WriteLine ("class MonoPackageManager_Resources {");
+				pkgmgr.WriteLine ("public class MonoPackageManager_Resources {");
 				pkgmgr.WriteLine ("\tpublic static final String[] Assemblies = new String[]{");
 
 				pkgmgr.WriteLine ("\t\t/* We need to ensure that \"{0}\" comes first in this list. */", mainFileName);
@@ -114,7 +108,7 @@ namespace Xamarin.Android.Tasks
 				pkgmgr.Flush ();
 
 				// Only copy to the real location if the contents actually changed
-				var dest = Path.GetFullPath (Path.Combine (OutputDirectory, "MonoPackageManager.java"));
+				var dest = Path.GetFullPath (Path.Combine (OutputDirectory, "MonoPackageManager_Resources.java"));
 
 				MonoAndroidHelper.CopyIfStreamChanged (stream, dest);
 			}