--- a/src/java-runtime/java/mono/android/MonoPackageManager.java
+++ b/src/java-runtime/java/mono/android/MonoPackageManager.java
@@ -106,4 +106,11 @@ public class MonoPackageManager {
 		return MonoPackageManager_Resources.ApiPackageName;
 	}
 }
+class MonoPackageManager_Resources {
+	public static final String[] Assemblies = new String[]{
+	};
+	public static final String[] Dependencies = new String[]{
+	};
+	public static final String ApiPackageName = "";
+}