--- a/src/java-runtime/java/mono/android/app/ApplicationRegistration.java
+++ b/src/java-runtime/java/mono/android/app/ApplicationRegistration.java
@@ -1,0 +1,9 @@
+package mono.android.app;
+
+public class ApplicationRegistration {
+
+	public static void registerApplications ()
+	{
+		// REGISTER_APPLICATION_AND_INSTRUMENTATION_CLASSES_HERE		
+	}
+}