--- a/src/java-runtime/java/mono/android/app/XamarinAndroidEnvironmentVariables.java
+++ b/src/java-runtime/java/mono/android/app/XamarinAndroidEnvironmentVariables.java
@@ -1,0 +1,9 @@
+package mono.android.app;
+
+public class XamarinAndroidEnvironmentVariables
+{
+	// Variables are specified the in "name", "value" pairs
+	public static final String[] Variables = new String[] {
+//@ENVVARS@
+	};
+}