… Xamarin.Android.sln
… src/java-runtime/java/mono/android/GCUserPeer.java
… src/java-runtime/java/mono/android/IGCUserPeer.java
… src/java-runtime/java/mono/android/MonkeyPatcher.java
… src/java-runtime/java/mono/android/ResourcePatcher.java
… src/java-runtime/java/mono/android/Runtime.java
… src/java-runtime/java/mono/android/Seppuku.java
… src/java-runtime/java/mono/android/app/NotifyTimeZoneChanges.java
… src/java-runtime/java/mono/android/debug/MultiDexLoader.java
… src/java-runtime/java/mono/android/incrementaldeployment/IncrementalClassLoader.java
… src/java-runtime/java/mono/android/incrementaldeployment/Placeholder.java
… src/java-runtime/java/mono/android/release/MultiDexLoader.java