@dellis1972 2019-01-04
This should be good to go. It will need changes in the close source part though. So we will probably need to bump in the same commit as the required fixes.