@jonathanpeppers 2019-01-10
I restarted the macOS build because it timed out.