@jonpryor 2019-01-11
@dellis1972: before we can merge, we need a decent commit message.  I've started writing one -- will provide in a later comment on this PR -- but as part of the message it would be nice to include the build time improvements.  *Presumably* by introducing the new `runtime.jar` file there will be fewer things to compile as part of build + packaging; what kind of improvements can we expect?