@jonpryor 2019-01-11
@dellis1972: Additionally, as we already create `mono.android.dex` files for the installer, should this PR should also create corresponding `runtime.dex` and `runtime_fastdev.dex` files?