@jonpryor 2019-01-11
Possibly silly suggestion? Instead of e.g. `runtime.jar` as the output file, should it instead be `java-runtime.jar`, sharing the name of the new project?  That might make it a bit easier to understand where files are coming from, but equally might not be necessary.