@jonpryor 2019-01-11
...Now that I think of it, will the fact that `MonoPackageManager.class` requires `MonoPackageManager_Resources.class` prevent `dx runtime.jar` from working, as `MonoPackageManager_Resources.class` doesn't exist?  I don't know.