@jonpryor 2019-01-11
Work-in-progress commit message:

```
A way to improve packaging times is to Do Less Work™.  There are a
few files that *every* project contains, that are present
*in this repo*, and don't really need to be built as part of the
project; they can instead be built and included with the installer,
thus reducing the need to run `javac` and `dx` during packaging.

These files include, but are not limited to:

  * `IncrementalClassLoader.java`
  * `MonkeyPatcher.java`
  * `MonoPackageManager.java`
  * `MultiDexLoader.java`
  * `NotifyTimeZoneChanges.java`
  * `ResourcePatcher.java`

Instead of storing these and other files as `@(EmbeddedResource)`s
within `Xamarin.Android.Build.Tasks.dll`, which then need to be
extracted and compiled as part of the app project, compile them into
`runtime.jar` and `runtime_fastdev.jar` files.

Add a new `src/java-runtime` project which creates the new
`runtime.jar` and `runtime_fastdev.jar` files.

A note about "circular dependencies" and `MonoPackageManager`:
previously, `MonoPackageManager.java` was stored as an
`@(EmbeddedResource)`, extracted, then *appended to* so that we could
generate the `MonoPackageManager_Resources` type, which is referenced
by `MonoPackageManager`.  However, `MonoPackageManager` is a type
we'd like to avoid compiling as part of the app project.  In order to
support this, we build `MonoPackageManager` against a "dummy"
`MonoPackageManager_Resources` type, then *delete*
`MonoPackageManager_Resources.class` before creating `runtime.jar`.

`MonoPackageManager_Resources` is stil generated at build time, and
must be compiled alongside other Java assets such as Java Callable
Wrappers.

TODO: Build time improvements?
```