@dellis1972 2019-02-23
Windows Failure is not related to this PR

```
NuGet.Protocol.Core.Types.FatalProtocolException: Unable to load the service index for source https://msazure.pkgs.visualstudio.com/_packaging/Dev/nuget/v3/index.json. ---> System.Net.Http.HttpRequestException: Response status code does not indicate success: 401 (Unauthorized).
```