@jonpryor 2019-03-15
Final (?) commit message:

```
Related: https://github.com/xamarin/designer/issues/1597

There are a few files that *every* project contains, that are present
*in this repo*, and don't really need to be built as part of any
other project; they can instead be built and included with the
installer, thus reducing the need to run `javac` and `dx` during
packaging.

These files include, but are not limited to:

  * `IncrementalClassLoader.java`
  * `MonkeyPatcher.java`
  * `MonoPackageManager.java`
  * `MultiDexLoader.java`
  * `NotifyTimeZoneChanges.java`
  * `ResourcePatcher.java`

Instead of storing these and other files as `@(EmbeddedResource)`s
within `Xamarin.Android.Build.Tasks.dll`, which then need to be
extracted and compiled as part of the app project, compile them into
`java_runtime.jar` and `java_runtime_fastdev.jar` files.

Add a new `src/java-runtime` project which creates the new
`java_runtime.jar` and `java_runtime_fastdev.jar` files.  It will
also create the `.dex` files for each of these.

A note about "circular dependencies" and `MonoPackageManager`:
previously, `MonoPackageManager.java` was stored as an
`@(EmbeddedResource)`, extracted, then *appended to* so that we could
generate the `MonoPackageManager_Resources` type, which is referenced
by `MonoPackageManager`.  However, `MonoPackageManager` is a type
we'd like to avoid compiling as part of the app project.  In order to
support this, we build `MonoPackageManager` against a "dummy"
`MonoPackageManager_Resources` type, then *delete*
`MonoPackageManager_Resources.class` before creating `java_runtime.jar`.

`MonoPackageManager_Resources` is stil generated at build time, and
must be compiled alongside other Java assets such as Java Callable
Wrappers.

There is also an issue with the `Dexes` fast deployment system.
On certain emulators the ART runtime complains if a `.class` is in
multiple `.dex` files. This results in the following error

	java.lang.IllegalAccessError: Class ref in pre-verified class resolved to unexpected implementation

and the app will fail to start. This was caused because the classes
`Runtime.class` and IGCUserPeer.class` ended up in both the final
`classes.dex` as well as the `mono.android.dex`.  This PR moves the
`Runtime.java` and IGCUserPeer.java` into the new `java_runtime.dex`
and out of `mono.android.dex` so we can avoid this issue.  The fast
deployment system will make use of the `java_runtime_fastdev.dex`
as the main `classes.dex` in a fast dev apk.  This means we only
include only the required files in the apk to get the app running.
All other files will be resolved from the fast deployment directory.

Using `java_runtime.jar` saves ~190ms from App builds, as the types
within `java_runtime.jar` don't need to be separately compiled.
These are the task execution times over 3 builds of the
`Xamarin.Forms-Performance-Integration` test app:

	[Before PR#2536]
	2578 ms  Javac                                      1 calls
	2474 ms  Javac                                      1 calls
	2928 ms  Javac                                      1 calls
	
	Avg: 2660
	
	[With `java_runtime.jar`]
	2497 ms  Javac                                      1 calls
	2500 ms  Javac                                      1 calls
	2424 ms  Javac                                      1 calls
	
	Avg: 2473

One other addition in this PR is the ability to change the name of a
`.dex` file within the Apk.  One of the problems we face is the first
`.dex` file in an apk MUST be called `classes.dex`.  Our current
`<BuildApk/>` system uses the item Filename for the apk entry.  This
works normally because we compile all the java code into
`classes.zip` and then into `classes.dex`.  However for the new fast
dev system we will just be using the `java_runtime.dex` file directly
so we need to ability to alter its name in the final apk.  Rather
than hardcoding this ability we just need to add the `%(ApkName)`
metadata to the appropriate item in MSBuild and the `<BuildApk/>`
task will now use that for the apk rather than the filename.
```