@jonpryor 2019-03-15
Now that e83ba0ddc159d15abcfac1a5a4bce5a53e165d3f has been merged, please update this PR to add `java_runtime.jar` & co. to `build-tools/installers/create-installers.targets`.